+++ CHANGELOG.md
@@ -3,6 +3,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- Disable the `big_image_size_threshold` feature to prevent scaling images. This breaks Stateless functionalities.
+- If on admin side, attachment URLs are filtered to point to the bucket.
+
+### Changed
+- Filter URLs only if not on admin side.
+
 ## [1.0.1] - 2017-10-13
 
 ### Added